+++ plugin-route-model/src/main/kotlin/com/linecorp/intellij/plugins/armeria/explorer/model/GrpcRoutePath.kt
@@ -1,0 +1,8 @@
+package com.linecorp.intellij.plugins.armeria.explorer.model
+
+object GrpcRoutePath {
+    fun path(
+        fqService: String,
+        methodName: String,
+    ): String = "/$fqService/$methodName"
+}